Split lsf test in two files.
